backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
praw==7.7.1
spacy==3.7.5
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
vaderSentiment==3.3.2
geopy==2.4.1
httpx==0.27.2
python-slugify==8.0.4
